+++ Natours_API/controllers/tourController.js
@@ -1,83 +1,112 @@
 /* eslint-disable prettier/prettier */
-const fs = require('fs');
+  //const fs = require('fs');
+const Tour = require('./../models/tourModel')
 
 
-
-const tours = JSON.parse (
-    fs.readFileSync(`${__dirname}/../dev-data/data/tours-simple.json`)
-    );
-    exports.checkID = (req,res,next,val) =>{
-        console.log(`Tour id is: ${val}`);
-        if(req.params.id*1>tours.length){
-            return res.status(404).json({
-                status: 'fail',
-                message: 'Invalid Id'
-            });
-        } 
-        next();
-    }; 
-    exports.checkBody = (req, res, next) => {
-        if(!req.body.name|| !req.body.price){
-            return res.status(400).json({
-                status: "fail",
-                message:"Missing name  or price"
-            })
-          }
-          next();  
-    }
 //getall 
-exports.getAllTours = (req,res) =>{
+exports.getAllTours = async (req,res) =>{
+
+    try{
+        const  tours = await Tour.find()
+
     res.status(200).json({
      status:"success",
-        requestedAt: req.requestTime,
+       
         results: tours.length,
         data: {
              tours
         }
      });
-    }
-//get
-exports.getTour = (req,res) =>{
+    
 
-    const id = req.params.id * 1;
-    const tour = tours.find(el=> el.id === id);
+    }catch(err){
+        res.status(404).json({
+            status: 'fail',
+            message: err
+        })
+
+    }
+}
+
+  
+//get
+exports.getTour = async (req,res) =>{
+
+    try{
+          const tour  = await Tour.findById(req.params.id)
+          res.status(200).json({
+            status:"success",
+         
+            data: {
+                 tour
+            }
+         });
+    }catch(err){
+        res.status(404).json({
+            status: 'fail',
+            message: err
+        })
+    }
+ 
+
+//     const tour = tours.find(el=> el.id === id);
    
   
-    res.status(200).json({
-    status:"success",
  
-    data: {
-         tour
-    }
- });
 }
 //post
-exports.createTour = (req,res) =>{
+exports.createTour = async (req,res) =>{
 
-    const newId = tours[tours.length-1].id+1;
-    // eslint-disable-next-line prefer-object-spread
-    const newTour =Object.assign({id:newId},req.body);
-    tours.push(newTour);
-    fs.writeFile(`${__dirname}/dev-data/data/tours-simple.json`,JSON.stringify(tours) ,  err =>{
-    res.status("201").json({
+    try{
+
+    
+    // const newTour = newTour({})
+    // newTour.save()
+
+   const newTour = await  Tour.create(req.body)
+
+    res.status(201).json({
         status:"success",
         data:{
             tour:newTour
         }
     });
-    });
-    }
-    // patch
-    exports.updateTour = (req,res) =>{
+}catch(err){
 
-   
-       res.status("200").json({
-           status: 'success',
-           data:{
-               tour: '<Updated tour here...'
-           }
-       });
-   }
+        res.status(400).json({
+            status: 'fail',
+            message: 'invalid data sent'
+        })
+
+}
+};
+    // patch
+    exports.updateTour = async (req,res) =>{
+
+        try{
+
+           const  tour = await Tour.findByIdAndUpdate(req.params.id, req.body, {
+               new: true,
+               runValidators: true
+        })
+
+            res.status("200").json({
+                status: 'success',
+                data:{
+                    tour
+                }
+            });
+                
+
+        }catch(err){
+
+            res.status(404).json({
+                status: 'fail',
+                message: err
+            });
+
+        }
+};
    //delete 
    exports.deleteTour = (req,res) =>{
