+++ Natours_API/server.js
@@ -20,38 +20,6 @@ mongoose
  // eslint-disable-next-line no-console
 //console.log(process.env);
 
-
-const  tourSchema = new mongoose.Schema({
-  name:{
-    type: String,
-    required: [true,'a tour  must have  a name'],
-    unique: true
-  },
-  rating: {
-    type: Number,
-    default: 4.5
-  },
-  price:{
-    type: Number,
-    required:[true, " A tour must have a price"]
-  }
-
-});
-const Tour = mongoose.model('Tour', tourSchema);
-
-const testTour = new Tour({
-  name:'the forset  hiker',
-  rating : 4.7,
-  price: 497
- });
-
-testTour.save().then(doc =>{
-  console.log(doc);
-
-}).catch(err =>{
-  console.log('Error:',err)
-})
-
 const port = process.env.PORT || 4000;
 app.listen(port, () => {
   // eslint-disable-next-line no-console
